rename secrets folder
