+++ ui/src/api/secrets_obsolete/mutate/deleteSecret.ts
