+++ ui/src/api/secrets_obsolete/mutate/updateSecret.ts
