+++ ui/src/api/secrets_obsolete/query/get.ts
