+++ ui/src/api/secrets_obsolete/schema.ts
